chatbot/requirements.txt
# 이 프로젝트가 직접 사용하는 라이브러리 목록
anthropic
psycopg2-binary
boto3
pydantic
google-cloud-aiplatform
google-auth
